pyproject.toml
[tool.poetry]
name = "nonebot_plugin_lagrange"
version = "1.1.0"
description = "一款便于管理 Lagrange.Onebot 的插件。"
license = "GPL3"
authors = ["Lonely-Sails"]
readme = "README.md"
homepage = "https://github.com/Lonely-Sails/nonebot-plugin-lagrange"
repository = "https://github.com/Lonely-Sails/nonebot-plugin-lagrange"
documentation = "https://github.com/Lonely-Sails/nonebot-plugin-lagrange#readme"
keywords = ["nonebot", "nonebot2", "lagrange.onebot", "lagrange", "onebot v11", "onebot"]

[tool.poetry.dependencies]
httpx = "^0.27.0"
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-plugin-localstore = "^0.7.0"
nonebot-adapter-onebot = { version = "^2.0.0-beta.1", optional = true }

[tool.poetry.extras]
onebot = ["nonebot-adapter-onebot"]
